+++ packages/price-reporter/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @renegade-fi/price-reporter
 
+## 0.0.19
+
+### Patch Changes
+
+- use renegade price topic
+- Updated dependencies
+  - @renegade-fi/token@0.0.19
+
 ## 0.0.18
 
 ### Patch Changes